refactor for multiframework
