+++ cola/algorithms/lanczos.py
@@ -2,7 +2,7 @@ from cola import SelfAdjoint
 from cola.fns import lazify
 from cola.ops import LinearOperator
 from cola.ops import Array
-from cola.ops import get_library_fns
+from cola.backends import get_library_fns
 from cola.utils import export
 from cola.utils.custom_autodiff import iterative_autograd
 import cola
