+++ cola/ops/operators.py
@@ -1,6 +1,6 @@
 from functools import reduce, partial
-from cola.ops.operator_base import LinearOperator
-from cola.ops.operator_base import Array, get_library_fns
+from cola.ops.operator_base import LinearOperator, Array
+from cola.backends import get_library_fns
 from cola.utils.dispatch import parametric
 import cola
 import numpy as np
