+++ setup.cfg
@@ -17,4 +17,4 @@ markers =
 [coverage:run]
 omit =
   cola/*_fns.py
-  cola/utils_test.py
+  cola/utils/test_utils.py
