+++ tests/algorithms/test_cg.py
@@ -6,9 +6,9 @@ from cola.algorithms.preconditioners import NystromPrecond
 from cola.algorithms.cg import run_batched_cg
 from cola.algorithms.cg import run_batched_tracking_cg
 from cola.algorithms.cg import run_cg
-from cola.utils_test import get_xnp, parametrize, relative_error
-from cola.utils_test import generate_spectrum, generate_pd_from_diag
-from cola.utils_test import generate_diagonals
+from cola.utils.test_utils import get_xnp, parametrize, relative_error
+from cola.utils.test_utils import generate_spectrum, generate_pd_from_diag
+from cola.utils.test_utils import generate_diagonals
 # from tests.algorithms.test_lanczos import construct_tridiagonal
