+++ tests/algorithms/test_lobpcg.py
@@ -1,8 +1,8 @@
 import numpy as np
 from cola.fns import lazify
 from cola.algorithms.lobpcg import lobpcg
-from cola.utils_test import get_xnp, parametrize, relative_error
-from cola.utils_test import generate_spectrum, generate_pd_from_diag
+from cola.utils.test_utils import get_xnp, parametrize, relative_error
+from cola.utils.test_utils import generate_spectrum, generate_pd_from_diag
 
 
 @parametrize(['torch', 'jax'])
