+++ tests/linalg/test_eig.py
@@ -5,8 +5,8 @@ from cola.ops import Identity
 from cola.ops import Triangular
 from cola.annotations import SelfAdjoint
 from cola.linalg.eigs import eig
-from cola.utils_test import get_xnp, parametrize, relative_error
-from cola.utils_test import generate_spectrum, generate_pd_from_diag
+from cola.utils.test_utils import get_xnp, parametrize, relative_error
+from cola.utils.test_utils import generate_spectrum, generate_pd_from_diag
 
 _tol = 1e-6
