+++ tests/linalg/test_solve.py
@@ -1,7 +1,7 @@
 import cola
 from cola.fns import lazify
-from cola.utils_test import get_xnp, parametrize, relative_error
-from cola.utils_test import generate_spectrum, generate_pd_from_diag
+from cola.utils.test_utils import get_xnp, parametrize, relative_error
+from cola.utils.test_utils import generate_spectrum, generate_pd_from_diag
 
 
 @parametrize(['torch', 'jax'])
