+++ tests/linalg/test_sqrt.py
@@ -3,8 +3,8 @@ from cola.fns import lazify
 from cola.linalg import sqrt
 from cola.ops import Diagonal
 from cola.annotations import SelfAdjoint
-from cola.utils_test import get_xnp, parametrize, relative_error
-from cola.utils_test import generate_spectrum, generate_pd_from_diag
+from cola.utils.test_utils import get_xnp, parametrize, relative_error
+from cola.utils.test_utils import generate_spectrum, generate_pd_from_diag
 
 _tol = 1e-6
