+++ tests/test_operators.py
@@ -15,7 +15,7 @@ from cola.ops import Sparse
 from cola.ops import Jacobian
 from cola.ops import LinearOperator
 from cola.algorithms.arnoldi import get_householder_vec
-from cola.utils_test import get_xnp, parametrize, relative_error
+from cola.utils.test_utils import get_xnp, parametrize, relative_error
 from linalg.operator_market import op_names, get_test_operator
 
 _tol = 1e-6
